tests/wpt/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-disabled.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
focused_element_id = null;

function OnFocus(event) {
  focused_element_id = event.target.id;
}
</script>

<selectmenu id="selectmenu" onfocus="OnFocus(event)" disabled>
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>

<script>
promise_test(async () => {
  const selectmenu = document.getElementById("selectmenu");
  selectmenu.focus();
  assert_equals(focused_element_id, null);
}, "Check that disabled <selectmenu> cannot be focused");
</script>
